configuracion basica webpack
