+++ index.html
@@ -1,5 +1,13 @@
-<html>
-    <head>
-        <script src="js/index.js"></script>
-    </head>
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Document</title>
+</head>
+<body>
+    
+    <script src="dist/bundle.js"></script>
+</body>
 </html>
